+++ gwt-material/src/main/java/gwt/material/design/client/MaterialDesignBase.java
@@ -24,7 +24,7 @@ import com.google.gwt.core.client.GWT;
 import com.google.gwt.core.client.ScriptInjector;
 import com.google.gwt.dom.client.StyleInjector;
 import com.google.gwt.resources.client.TextResource;
-import gwt.material.design.client.base.pwa.PwaManager;
+import gwt.material.design.client.pwa.PwaManager;
 import gwt.material.design.client.resources.MaterialResources;
 
 import java.util.ArrayList;
